+++ src/components/ErrorBoundary/index.tsx
@@ -40,4 +40,4 @@ class ErrorBoundary extends React.Component<
   }
 }
 
-export default ErrorBoundary;
+export { ErrorBoundary };
