+++ src/components/index.ts
@@ -1,5 +1,6 @@
 import { CharacterList } from './CharacterList';
-import Loader from './Loader/Loader';
+import { ErrorBoundary } from './ErrorBoundary';
+import { Loader } from './Loader';
 import { Search } from './Search';
 
-export { CharacterList, Search, Loader };
+export { CharacterList, Search, Loader, ErrorBoundary };
