+++ src/main.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
+import { ErrorBoundary } from './components/';
+import { SearchPage } from './pages/index.ts';
 import './index.css';
-import SearchPage from './SearchPage.tsx';
-import ErrorBoundary from './components/ErrorBoundary/index.tsx';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <ErrorBoundary>
