+++ src/pages/SearchPage/index.ts
@@ -1,0 +1,1 @@
+export { default as SearchPage } from './SearchPage';
