+++ src/pages/SearchPage/style.module.scss
