+++ src/pages/index.ts
@@ -1,0 +1,3 @@
+import { SearchPage } from './SearchPage';
+
+export { SearchPage };
